back-end/volumes/app/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cleanup==6.0.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
persiantools==3.0.0
Pillow==9.2.0
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
